requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
anyio==4.7.0
appier==1.34.6
astroplan==0.10.1
astropy==7.1.0
astropy-iers-data==0.2025.5.26.0.37.21
astroquery==0.4.7
asyncclick==8.1.7.2
attrs==24.2.0
audioop-lts==0.2.1
audioread==3.0.1
beautifulsoup4==4.12.3
bidict==0.23.1
blurhash==1.1.4
bs4==0.0.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
decorator==5.1.1
dropbox==12.0.2
dropbox-api==0.2.6
fonttools==4.55.2
frozenlist==1.5.0
graphviz==0.20.3
html5lib==1.1
idna==3.10
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
joblib==1.5.0
july==0.1.3
keyring==25.5.0
kiwisolver==1.4.7
lark==1.2.2
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
mashumaro==3.15
Mastodon.py==1.8.1
matplotlib==3.9.3
more-itertools==10.5.0
msgpack==1.1.0
multidict==6.1.0
numba==0.61.2
numpy
opencv-python==4.10.0.84
packaging==24.2
paho-mqtt==2.1.0
pandas==2.3.0
pillow==11.0.0
platformdirs==4.3.7
ply==3.11
pooch==1.8.2
propcache==0.2.1
PyAstronomy==0.22.0
PyAudio==0.2.14
pycparser==2.22
pyerfa==2.0.1.5
pyparsing==3.2.0
pysolar==0.11
python-dateutil==2.9.0.post0
python-kasa==0.8.1
python-magic==0.4.27
python-weather==2.0.7
pytz==2024.2
pyvo==1.6
PyYAML==6.0.2
quantities==0.16.1
requests==2.32.3
scikit-learn==1.6.1
scikit-image
scipy==1.14.1
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.13.1
soupsieve==2.6
soxr==0.5.0.post1
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
stone==3.3.1
threadpoolctl==3.6.0
tplink-cloud-api==4.2.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.2.3
webencodings==0.5.1
yarl==1.18.3
speedtest-cli

